# files/api.yml
openapi: 3.0.3
info:
  title: GO Money Tracker
  description: |-
    This is an API for tracking expenses
  contact:
    email: angelbachev@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
  - url: http://104.248.29.228:8089
  - url: http://localhost:8089
...
paths:
  /api/user/settings:
    put:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Update user settings
      operationId: updateUserSettings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                defaultAccountId:
                  type: number
                  format: int64
                theme: 
                  type: string
                  enum:
                    - default
                    - light
                    - dark
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'          
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    get:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Get user settings
      operationId: getUserSettings
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSettings'
  /api/accounts:
    post:
      security:
        - bearerAuth: []
      tags:
        - Accounts
      summary: Create a new account
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description: 
                  type: string
                currencyCode:
                  type: string
                  format: currency
                  example: BGN
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'          
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    get:
      security:
        - bearerAuth: []
      tags:
        - Accounts
      summary: List accounts
      operationId: listAccounts
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
  /api/accounts/{accountId}:
    parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      security:
        - bearerAuth: []
      tags:
        - Accounts
      summary: Update account
      operationId: updateAccount
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description: 
                  type: string
                currencyCode:
                  type: string
                  format: currency
                  example: BGN
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'          
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '404':
          description: Not found     
    delete:
      security:
        - bearerAuth: []
      tags:
        - Accounts
      summary: Delete account
      operationId: deleteAccount
      parameters:
        - name: force
          in: query
          description: If param is present and its value is not 0 all categories and expenses in all categories will be deleted. Otherwise if account is not empty an error will be returned
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Successful operation
        '404':
          description: Not found     
  /api/accounts/{accountId}/categories:
    parameters:
      - name: accountId
        in: path
        description: ID of account for which we want to create category
        required: true
        schema:
          type: integer
          format: int64
    post:
      security:
        - bearerAuth: []
      tags:
        - Categories
      summary: Create category
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description: 
                  type: string
                parentId:
                  type: integer
                  format: int64
                  nullable: true
                icon:
                  type: string
                
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'          
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    get:
      security:
        - bearerAuth: []
      tags:
        - Categories
      summary: Get category tree
      operationId: getCategoryTree
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryTree'        
  /api/accounts/{accountId}/categories/{categoryId}:
    parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: categoryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      security:
        - bearerAuth: []
      tags:
        - Categories
      summary: Get category tree for specific category
      operationId: getCategoryTreeForSpecifiCategory
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryTree'       
    delete:
      security:
        - bearerAuth: []
      tags:
        - Categories
      summary: Delete category
      operationId: deleteCategory
      parameters:
        - name: force
          in: query
          description: If param is present and its value is not 0 all subcategories and expenses in category and its subcategories will be deleted. Otherwise if category is not empty an error will be returned
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Successful operation
        '404':
          description: Not found      
    put:
      security:
        - bearerAuth: []
      tags:
        - Categories
      summary: Update category
      operationId: updateCategory
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - parentId
                - icon
              properties:
                name:
                  type: string
                description: 
                  type: string
                parentId:
                  type: integer
                  format: int64
                  nullable: true
                icon:
                  type: string
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'          
        '400':
          description: Invalid input
        '422':
          description: Validation exception
...
  /api/accounts/{accountId}/expenses/{expenseId}:
    parameters:
      - name: accountId
        in: path
        description: ID of account for which we want to create expense
        required: true
        schema:
          type: integer
          format: int64  
      - name: expenseId
        in: path
        description: ID of the expense to be deleted
        required: true
        schema:
          type: integer
          format: int64
    delete:
      security:
        - bearerAuth: []
      tags:
        - Expenses
      summary: Delete expense
      operationId: deleteExpense
      responses:
        '204':
          description: Successful operation
        '404':
          description: Not found
    put:
      security:
        - bearerAuth: []
      tags:
        - Expenses
      summary: Update expense
      operationId: updateExpense
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - categoryId
                - description
                - amount
                - date
              properties:
                categoryId:
                  type: integer
                  format: int64
                description: 
                  type: string
                amount:
                  type: integer
                  format: int64
                  description: Amount in minor units eg. cents
                  example: 123.12
                date:
                  type: string
                  format: date-time
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'          
        '400':
          description: Invalid input
        '422':
          description: Validation exception
  /api/accounts/{accountId}/expenses/import:
    post:
      tags:
        - Expenses
      summary: Import expenses
      operationId: importExpenses
      parameters:
        - name: accountId
          in: path
          description: ID of account for which we want to create expense
          required: true
          schema:
            type: integer
            format: int64  
      requestBody:
        description: Format of the file is Date (dd.mm.YYYY), CategoryName, Amount (25,14 лв.), Description
        content:
          text/csv:
            schema:
              type: string
              format: binary
      responses:
        '204':
          description: successful operation
      security:
        - bearerAuth: []
  /api/users:
    post:
      tags:
        - Users
      summary: Create user (Register)
      operationId: register
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Account:
      required:
        - id
        - name
        - description
        - currencyCode
        - createdAt
        - updatedAt
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        currencyCode:
          type: string
          format: currency
          example: BGN
        createdAt:
          type: string
          format: date-time
          example: '2024-04-20T16:42:01Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-04-20T16:42:01Z'
...
    UserSettings:
      required:
        - defaultAccountId
        - theme
        - createdAt
        - updatedAt
      type: object
      properties:
        defaultAccountId:
          type: integer
          format: int64
        theme:
          type: string
        createdAt:
          type: string
          format: date-time
          example: '2024-04-20T16:42:01Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-04-20T16:42:01Z'